TrueNAS Rust NSS Library
A complete Rust implementation of NSS (Name Service Switch) functionality
from glibc. The primary feature that this library introduces is the
ability to specify which NSS module from which to retrieve the users or
groups if desired, and the return value includes information about the
NSS module that resolved it. This is used by the TrueNAS middleware and
other services to determine the origin of the account (whether from
a directory service or a local account).